rebased to fresh 9font APE with bitfield patch included, reapplied Lufias patch

+++ sys/src/ape/lib/bsd/gethostname.c
@@ -6,7 +6,7 @@
 #include <string.h>
 
 int
-gethostname(char *name, size_t namelen)
+gethostname(char *name, int namelen)
 {
 	int n, fd;
 	char buf[128];
